MULE-11603: Adding new EL micro-benchmarks to compare vs MEL
