+++ tests/performance/src/test/java/org/mule/test/core/el/ExpressionBenchmarkAssertionTestCase.java
@@ -6,11 +6,10 @@
  */
 package org.mule.test.core.el;
 
-import static java.util.Collections.singletonMap;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
 import org.mule.AbstractBenchmarkAssertionTestCase;
-import org.mule.el.SimpleExpressionBenchmark;
+import org.mule.el.ExpressionBenchmark;
 
 import org.junit.Test;
 
@@ -20,19 +19,13 @@ public class ExpressionBenchmarkAssertionTestCase extends AbstractBenchmarkAsser
 
   @Test
   public void evaluatePayloadMEL() {
-    runAndAssertBenchmark(SimpleExpressionBenchmark.class, "evaluatePayload", 1, singletonMap(EXPRESSION_PARAM,
-                                                                                              new String[] {
-                                                                                                  "mel:payload"}),
-                          700, NANOSECONDS, 2000);
+    runAndAssertBenchmark(ExpressionBenchmark.class, "melPayload", 700, NANOSECONDS, 2000);
   }
 
   @Test
   public void evaluatePayloadDW() {
     // TODO MULE-11971 DW numbers are significantly below MEL currently but by adding assertions we avoid any regressions
-    runAndAssertBenchmark(SimpleExpressionBenchmark.class, "evaluatePayload", 1, singletonMap(EXPRESSION_PARAM,
-                                                                                              new String[] {
-                                                                                                  "payload"}),
-                          5000, NANOSECONDS, 7000);
+    runAndAssertBenchmark(ExpressionBenchmark.class, "dwPayload", 5000, NANOSECONDS, 7000);
   }
 
 }
